+++ python/requirements.txt
@@ -1,13 +1,5 @@
 beautifulsoup4==4.12.3
 requests==2.32.3
-flask==3.0.3
 openai==1.58.1
 python-dotenv==1.0.1
-sentence-transformers==3.1.1
-numpy==2.1.1
-scikit-learn==1.5.2
-flask-cors==5.0.0
 fpdf2==2.8.2 
-pymongo==4.11.1
-bcrypt==4.2.1
-PyJWT==2.10.1
